+++ src/main/java/pl/asie/computronics/gui/providers/GuiProviderCipher.java
@@ -1,5 +1,7 @@
 package pl.asie.computronics.gui.providers;
 
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.world.World;
 import pl.asie.computronics.gui.GuiCipherBlock;
@@ -14,6 +16,7 @@ import pl.asie.lib.gui.managed.LegacyGuiProvider;
  */
 public class GuiProviderCipher extends LegacyGuiProvider {
 	@Override
+	@SideOnly(Side.CLIENT)
 	protected GuiBase makeGuiBase(int i, EntityPlayer entityPlayer, World world, int x, int y, int z, TileEntityBase tile) {
 		return new GuiCipherBlock(makeContainerBase(guiID, entityPlayer, world, x, y, z, tile));
 	}